Extract "h7_handler" into separate module.
