+++ include/h7_handler.h
@@ -1,0 +1,34 @@
+/*
+ * Firmware for the Portenta X8 STM32H747AIIX/Cortex-M7 core.
+ * Copyright (C) 2022 Arduino (http://www.arduino.cc/)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#ifndef PORTENTAX8_STM32H7_FW_H7_HANDLER_H
+#define PORTENTAX8_STM32H7_FW_H7_HANDLER_H
+
+/**************************************************************************************
+ * INCLUDE
+ **************************************************************************************/
+
+#include <stdint.h>
+
+/**************************************************************************************
+ * FUNCTION DECLARATION
+ **************************************************************************************/
+
+int h7_handler(uint8_t opcode, uint8_t *data, uint16_t size);
+
+#endif /* PORTENTAX8_STM32H7_FW_H7_HANDLER_H */
